Renamed mycomponents to components
