+++ components/chakra-next-image.tsx
