+++ components/mdx/components/index.tsx
@@ -8,7 +8,7 @@ import {
 } from '@chakra-ui/react';
 import { Code, preToCodeBlock } from './code';
 import { CodePen, CodeSandbox, Tweet } from 'mdx-embed';
-import { ChakraLink } from 'mycomponents/chakra-link';
+import { ChakraLink } from 'components/chakra-link';
 
 // Components provided by Chakra UI https://github.com/chakra-ui/chakra-ui/blob/main/website/src/components/mdx-components.tsx
